Module node_template::chain_spec

Enums

Alternative
The chain specification option. This is expected to come in from the CLI and is little more than one of a number of alternatives which can easily be converted from a string (--chain=...) into a ChainSpec.

Functions

get_account_id_from_seed
Helper function to generate an account ID from seed

get_authority_keys_from_seed
Helper function to generate an authority key for Aura

get_from_seed
Helper function to generate a crypto pair from seed

load_spec

testnet_genesis

Type Definitions

AccountPublic

ChainSpec
Specialized ChainSpec. This is a specialization of the general Substrate ChainSpec type.